Shape Movement
Got the shape to move(Need to fix the screen)

+++ core/src/com/rabbitstudios/getout/GetOut.java
@@ -12,7 +12,7 @@ public class GetOut extends Game {
 	public SpriteBatch batch;
     public static final int V_WIDTH = 400;
 	public static final int V_HEIGHT = 208;
-    public static final int PPM = 100;
+    public static final float PPM = 100;
 	@Override
 	public void create () {
 		batch = new SpriteBatch();
